Started development ServiceOrderRepository

+++ FIAPOficina.Api/Models/Services/Responses/ServiceResponse.cs
@@ -5,6 +5,6 @@
         Guid Id,
         string Name,
         string Description,
-        double Value
+        decimal Value
     );
 }
