+++ FIAPOficina.Application/Services/Commands/CreateService/CreateServiceCommand.cs
@@ -1,4 +1,4 @@
 ﻿namespace FIAPOficina.Application.Services.Commands.CreateService
 {
-    public record CreateServiceCommand(string Name, string Description, double Value);
+    public record CreateServiceCommand(string Name, string Description, decimal Value);
 }
