+++ FIAPOficina.Domain/Services/Entities/Service.cs
@@ -5,9 +5,9 @@
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public double Value { get; set; }
+        public decimal Value { get; set; }
 
-        public Service(string name, string description, double value, Guid? id = null)
+        public Service(string name, string description, decimal value, Guid? id = null)
         {
             Name = name;
             Description = description;
